refactor(action): fold ClearAppDataIos into ClearAppData platform switch
Merge the Android and iOS clear-app-data commands into a single
ClearAppData class that dispatches on device.platform via a switch, with
private executeAndroid/executeIos methods (issue #5461). Replaces the
call-site ternary in AppCleanupService and the ClearAppDataIos
construction in LaunchApp with a single ClearAppData construction; the
iOS-only constructor guard is dropped in favor of the platform switch.

Behavior-preserving. iOS coverage moves into ClearAppData.test.ts.
