+++ src/server/AppCleanupService.ts
@@ -1,6 +1,5 @@
 import { BootedDevice, ClearAppDataResult, TerminateAppResult } from "../models";
 import { ClearAppData } from "../features/action/ClearAppData";
-import { ClearAppDataIos } from "../features/action/ClearAppDataIos";
 import { TerminateApp } from "../features/action/TerminateApp";
 import { Logger, logger } from "../utils/logger";
 
@@ -40,8 +39,7 @@ export class DefaultAppCleanupService implements AppCleanupService {
 
   constructor(dependencies: AppCleanupDependencies = {}) {
     this.createClearAppData =
-      dependencies.createClearAppData ?? ((device: BootedDevice) =>
-        device.platform === "ios" ? new ClearAppDataIos(device) : new ClearAppData(device));
+      dependencies.createClearAppData ?? ((device: BootedDevice) => new ClearAppData(device));
     this.createTerminateApp =
       dependencies.createTerminateApp ?? ((device: BootedDevice) => new TerminateApp(device));
     this.log = dependencies.logger ?? logger;
